Fix #3, use sizeof() for header sizes.
sizeof() will keep this tool in sync if the size of the CFE
file or table header should ever change.

Note this also implicitly restricts this tool to the CFE build
environment where these headers are available.  As a result
this will no longer be buildable as a fully standalone/external
tool.